Moved to float based render buffers

+++ application/view/indexedMesh.cpp
@@ -3,7 +3,7 @@
 #include "../util/log.h"
 
 IndexedMesh::IndexedMesh(const Shape& shape) : AbstractMesh(), vertexCount(shape.vertexCount), triangleCount(shape.triangleCount) {
-	double* vertices = new double[shape.vertexCount * 3];
+	float* vertices = new float[shape.vertexCount * 3];
 	for(int i = 0; i < vertexCount; i++) {
 		Vec3f vertex = shape[i];
 		vertices[i * 3] = vertex.x;
@@ -12,30 +12,30 @@ IndexedMesh::IndexedMesh(const Shape& shape) : AbstractMesh(), vertexCount(shape
 	}
 
 	vertexBuffer = new VertexBuffer(vertices, vertexCount * 3);
-	normalBuffer = new VertexBuffer(reinterpret_cast<double const *>(shape.normals.get()), vertexCount * 3);
-	uvBuffer = new VertexBuffer(reinterpret_cast<double const *>(shape.uvs.get()), vertexCount * 2);
+	normalBuffer = new VertexBuffer(reinterpret_cast<float const *>(shape.normals.get()), vertexCount * 3);
+	uvBuffer = new VertexBuffer(reinterpret_cast<float const *>(shape.uvs.get()), vertexCount * 2);
 
 	indexBuffer = new IndexBuffer(reinterpret_cast<unsigned int const *>(shape.triangles), triangleCount * 3);
 
-	vertexBufferLayout.push<double>(3);
-	normalBufferLayout.push<double>(3);
-	uvBufferLayout.push<double>(2);
+	vertexBufferLayout.push<float>(3);
+	normalBufferLayout.push<float>(3);
+	uvBufferLayout.push<float>(2);
 
 	vertexArray->addBuffer(*vertexBuffer, vertexBufferLayout);
 	vertexArray->addBuffer(*normalBuffer, normalBufferLayout);
 	vertexArray->addBuffer(*uvBuffer, uvBufferLayout);
 }
 
-IndexedMesh::IndexedMesh(const double* vertices, const double* normals, const double* uvs, const unsigned int* indices, const int vertexCount, const int triangleCount) : AbstractMesh(), vertexCount(vertexCount), triangleCount(triangleCount) {
+IndexedMesh::IndexedMesh(const float* vertices, const float* normals, const float* uvs, const unsigned int* indices, const int vertexCount, const int triangleCount) : AbstractMesh(), vertexCount(vertexCount), triangleCount(triangleCount) {
 	vertexBuffer = new VertexBuffer(vertices, vertexCount * 3);
 	normalBuffer = new VertexBuffer(normals, vertexCount * 3);
 	uvBuffer = new VertexBuffer(uvs, vertexCount * 2);
 
 	indexBuffer = new IndexBuffer(indices, triangleCount * 3);
 
-	vertexBufferLayout.push<double>(3);
-	normalBufferLayout.push<double>(3);
-	uvBufferLayout.push<double>(2);
+	vertexBufferLayout.push<float>(3);
+	normalBufferLayout.push<float>(3);
+	uvBufferLayout.push<float>(2);
 
 	vertexArray->addBuffer(*vertexBuffer, vertexBufferLayout);
 	vertexArray->addBuffer(*normalBuffer, normalBufferLayout);
